Move Index-, Shader- and Vertex Buffer to their own folder
